convert sheet directly to bmp
