+++ handwrite/__init__.py
@@ -1,4 +1,4 @@
-from handwrite.sheettopng import SHEETtoPNG
-from handwrite.pngtosvg import PNGtoSVG
+from handwrite.sheettobmp import SHEETtoBMP
+from handwrite.bmptosvg import BMPtoSVG
 from handwrite.svgtottf import SVGtoTTF
 from handwrite.cli import converters
